+++ dev/app/wrap-with-hits.js
@@ -1,15 +1,43 @@
 /* eslint-disable import/default */
-import instantsearch from '../../index.js';
+import { action } from 'dev-novel';
 
+import instantsearch from '../../index.js';
 import item from './templates/item.html';
 import empty from './templates/no-results.html';
 
-export default (initWidget, opts = {}) => container => {
+export default (initWidget, instantSearchConfig = {}) => container => {
+  const {
+    appId = 'latency',
+    apiKey = '6be0576ff61c053d5f9a3225e2a90f76',
+    indexName = 'instant_search',
+    searchParameters = {},
+    ...otherInstantSearchConfig
+  } = instantSearchConfig;
+
+  const urlLogger = action('[URL sync] pushstate: query string');
   window.search = instantsearch({
-    appId: 'latency',
-    apiKey: '6be0576ff61c053d5f9a3225e2a90f76',
-    indexName: 'instant_search',
-    searchParameters: { hitsPerPage: 3, ...(opts.searchParameters || {}) },
+    appId,
+    apiKey,
+    indexName,
+    searchParameters: {
+      hitsPerPage: 3,
+      ...searchParameters,
+    },
+    urlSync: {
+      urlUtils: {
+        onpopstate() {},
+        pushState(qs) {
+          urlLogger(this.createURL(qs));
+        },
+        createURL(qs) {
+          return qs;
+        },
+        readUrl() {
+          return '';
+        },
+      },
+    },
+    ...otherInstantSearchConfig,
   });
 
   container.innerHTML = `
@@ -48,5 +76,6 @@ export default (initWidget, opts = {}) => container => {
   );
 
   initWidget(window.document.getElementById('widget-display'));
+
   window.search.start();
 };