+++ src/components/Slider/Slider.js
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types';
 
 import React, { Component } from 'preact-compat';
 
-import Rheostat from 'rheostat';
+import Rheostat from 'preact-rheostat';
 import cx from 'classnames';
 
 import Pit from './Pit.js';
@@ -48,9 +48,7 @@ export class RawSlider extends Component {
       min,
       ...times(steps - 1, step => min + stepsLength * (step + 1)),
       max,
-    ]
-      // bug with `key={ 0 }` and preact, see https://github.com/developit/preact/issues/642
-      .map(pitPoint => (pitPoint === 0 ? 0.000001 : pitPoint));
+    ];
 
     return pitPoints;
   }